change scorer weight type from int to float (#2207)
* change scorer weight type from int to float

this is a mandatory change to allow adaptive configuration of scorers weights, since the recalculated weights
of scorers may result in floating numbers and not only integers.


* fix config loader unit tests


---------

+++ apix/config/v1alpha1/endpointpickerconfig_types.go
@@ -126,13 +126,13 @@ type SchedulingPlugin struct {
 
 	// +optional
 	// Weight is the weight fo be used if this plugin is a Scorer.
-	Weight *int `json:"weight"`
+	Weight *float64 `json:"weight"`
 }
 
 func (sp SchedulingPlugin) String() string {
 	var weight string
 	if sp.Weight != nil {
-		weight = fmt.Sprintf(", Weight: %d", *sp.Weight)
+		weight = fmt.Sprintf(", Weight: %f", *sp.Weight)
 	}
 	return fmt.Sprintf("{PluginRef: %s%s}", sp.PluginRef, weight)
 }
